+++ jarn/mkrelease/tests/test_setuptools.py
@@ -270,6 +270,99 @@ testpackage.egg-info/top_level.txt""")
         self.assertEqual(contains(archive, 'git_only.py'), True)
 
 
+class NoneTests(GitSetup):
+
+    def testNoneSdistPy(self):
+        st = Setuptools(Process(quiet=True, env=get_env()))
+        archive = st.run_dist(self.packagedir, [], 'sdist', ['--formats=zip'], ff='none')
+        self.assertEqual(contains(archive, 'git_only.py'), True)
+
+    def testNoneSdistTxt(self):
+        st = Setuptools(Process(quiet=True, env=get_env()))
+        archive = st.run_dist(self.packagedir, [], 'sdist', ['--formats=zip'], ff='none')
+        self.assertEqual(contains(archive, 'git_only.txt'), False)
+
+    def testNoneSdistC(self):
+        st = Setuptools(Process(quiet=True, env=get_env()))
+        archive = st.run_dist(self.packagedir, [], 'sdist', ['--formats=zip'], ff='none')
+        self.assertEqual(contains(archive, 'git_only.c'), False)
+
+    def testNoneMetaFile(self):
+        st = Setuptools(Process(quiet=True, env=get_env()))
+        archive = st.run_dist(self.packagedir, [], 'sdist', ['--formats=zip'], ff='none')
+        self.assertEqual(contains(archive, '.gitignore'), False)
+
+    def testNoneManifest(self):
+        st = Setuptools(Process(quiet=True, env=get_env()))
+        archive = st.run_dist(self.packagedir, [], 'sdist', ['--formats=zip'], ff='none')
+        self.assertEqual(get_manifest(archive), """\
+README.txt
+setup.py
+testpackage/__init__.py
+testpackage/git_only.py
+testpackage.egg-info/PKG-INFO
+testpackage.egg-info/SOURCES.txt
+testpackage.egg-info/dependency_links.txt
+testpackage.egg-info/not-zip-safe
+testpackage.egg-info/requires.txt
+testpackage.egg-info/top_level.txt""")
+
+    def testRemoveSetupPyc(self):
+        st = Setuptools(Process(quiet=True, env=get_env()))
+        st.run_dist(self.packagedir, [], 'sdist', ['--formats=zip'], ff='none')
+        self.assertFalse(isfile(join(self.packagedir, 'setup.pyc')))
+
+
+class ManifestInTests(GitSetup):
+
+    def setUp(self):
+        GitSetup.setUp(self)
+        self.manifest_in = join(self.packagedir, 'MANIFEST.in')
+        self.mkfile(self.manifest_in, 'recursive-include testpackage *.c')
+
+    def testManifestSdistPy(self):
+        st = Setuptools(Process(quiet=True, env=get_env()))
+        archive = st.run_dist(self.packagedir, [], 'sdist', ['--formats=zip'], ff='none')
+        self.assertEqual(contains(archive, 'git_only.py'), True)
+
+    def testManifestSdistTxt(self):
+        st = Setuptools(Process(quiet=True, env=get_env()))
+        archive = st.run_dist(self.packagedir, [], 'sdist', ['--formats=zip'], ff='none')
+        self.assertEqual(contains(archive, 'git_only.txt'), False)
+
+    def testManifestSdistC(self):
+        st = Setuptools(Process(quiet=True, env=get_env()))
+        archive = st.run_dist(self.packagedir, [], 'sdist', ['--formats=zip'], ff='none')
+        self.assertEqual(contains(archive, 'git_only.c'), True)
+
+    def testManifestMetaFile(self):
+        st = Setuptools(Process(quiet=True, env=get_env()))
+        archive = st.run_dist(self.packagedir, [], 'sdist', ['--formats=zip'], ff='none')
+        self.assertEqual(contains(archive, '.gitignore'), False)
+
+    def testManifestManifest(self):
+        st = Setuptools(Process(quiet=True, env=get_env()))
+        archive = st.run_dist(self.packagedir, [], 'sdist', ['--formats=zip'], ff='none')
+        self.assertEqual(get_manifest(archive), """\
+MANIFEST.in
+README.txt
+setup.py
+testpackage/__init__.py
+testpackage/git_only.c
+testpackage/git_only.py
+testpackage.egg-info/PKG-INFO
+testpackage.egg-info/SOURCES.txt
+testpackage.egg-info/dependency_links.txt
+testpackage.egg-info/not-zip-safe
+testpackage.egg-info/requires.txt
+testpackage.egg-info/top_level.txt""")
+
+    def testRemoveSetupPyc(self):
+        st = Setuptools(Process(quiet=True, env=get_env()))
+        st.run_dist(self.packagedir, [], 'sdist', ['--formats=zip'], ff='none')
+        self.assertFalse(isfile(join(self.packagedir, 'setup.pyc')))
+
+
 def test_suite():
     return unittest.defaultTestLoader.loadTestsFromName(__name__)